// Backend/Store/src/homepage/home-layout.types.ts
export enum HomeSectionType {
  HERO_CAROUSEL = 'HERO_CAROUSEL',
  CATEGORY_STRIP = 'CATEGORY_STRIP',
  PRODUCT_CAROUSEL = 'PRODUCT_CAROUSEL',
  BRAND_STRIP = 'BRAND_STRIP',
  VALUE_PROPS = 'VALUE_PROPS',
  CUSTOM_HTML = 'CUSTOM_HTML',
}

export enum HomeSectionItemType {
  BANNER = 'BANNER',
  CATEGORY = 'CATEGORY',
  BRAND = 'BRAND',
  PRODUCT = 'PRODUCT',
  LINK = 'LINK',
}